fix(#105): 검색 결과 내 무한 스크롤 동작시 좋아요 순 정렬 유지
