+++ src/pages/search/category.tsx
@@ -1,4 +1,3 @@
-import { useEffect, useState } from 'react';
 import { GetServerSidePropsContext } from 'next';
 import { dehydrate } from '@tanstack/react-query';
 import SEO from '@/components/atoms/SEO';
@@ -9,6 +8,7 @@ import SortDropdown from '@/components/molecules/SortDropdown';
 import prefetchQuery from '@/services/prefetchQuery';
 import { getBookReviewListByCategoryInfinityQuery } from '@/services/queries/bookReview';
 import useInfinityBookReviewListByCategory from '@/hooks/services/infinityQueries/useInfinityBookReviewListByCategory';
+import useSortedBookReviewList from '@/hooks/useSortedBookReviewList';
 import { BookTitle } from '@/types/features/book';
 import { Category } from '@/types/features/category';
 import Route from '@/constants/routes';
@@ -17,20 +17,11 @@ const SearchResultPage = ({ category }: { category: Category }) => {
   const { bookReviewList: initBookReviewList, refetchBookReviewList } =
     useInfinityBookReviewListByCategory(category);
 
-  const [bookReviewList, setBookReviewList] = useState(initBookReviewList);
-
-  useEffect(() => {
-    setBookReviewList(initBookReviewList);
-  }, [initBookReviewList]);
-
-  const handleClickLatestSortButton = () => {
-    setBookReviewList(initBookReviewList);
-  };
-
-  const handleClickLikeSortButton = () => {
-    const list = [...bookReviewList];
-    setBookReviewList(list.sort((a, b) => b.likeCount - a.likeCount));
-  };
+  const {
+    bookReviewList,
+    handleClickLatestSortButton,
+    handleClickLikeSortButton,
+  } = useSortedBookReviewList(initBookReviewList);
 
   return (
     <>
